// src/app/api/models/event-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EventDto {
  city?: string;
  country?: string;
  displayName?: string;
  endDate?: string;
  eventId?: number;
  hasResults?: boolean;
  id?: number;
  infoLink?: string;
  isChampionship: boolean;
  isSwisstour: boolean;
  name?: string;
  numberPlayers?: number;
  points: number;
  purse?: number;
  registrationLink?: string;
  registrationStart?: string;
  startDate?: string;
  swisstourType?: string;
  tier?: string;
  year?: number;
}